fix: Added newly nested attributes within the ZPA  and  models.
